+++ template/src/index.html.jsx
@@ -1,3 +1,7 @@
+require('./images/favicon.png');
+require('./manifest.json');
+require('./robots.txt');
+
 import * as React from "react";
 import ReactDOMServer from "react-dom/server";
 import {Provider} from "react-redux";
@@ -5,10 +9,14 @@ import {StaticRouter as Router} from "react-router";
 
 import {store} from "./js/data/store";
 
+const structuredData = {
+
+};
+
 const index = ReactDOMServer.renderToStaticMarkup(
-  <html lang="en">
+  <html>
     <head>
-      <title></title>
+      <title>{env.title}</title>
       <meta httpEquiv="X-UA-Compatible" content="IE=edge"/>
       <meta charSet="utf8"/>
       <meta name="description" content={env.description}/>
@@ -16,18 +24,37 @@ const index = ReactDOMServer.renderToStaticMarkup(
       <meta name="author" content={env.author}/>
       <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
       <meta name="apple-mobile-web-app-capable" content="yes"/>
-      <meta name="theme-color" content="#8b0000"/>
+      <meta name="theme-color" content="#000000"/>
       <link rel="icon" href="/images/favicon.png" sizes="32x32"/>
       <link rel="shortcut icon" href="images/favicon.png" sizes="196x196"/>
       <link rel="manifest" href="/manifest.json"/>
 
-      {env.environment === "production" && <style dangerouslySetInnerHTML={{__html: require("./css/main.scss").toString()}}/>}
-      {env.environment === "development" && <script defer={true} type="text/javascript" src="webpack-bootstrap.js"></script>}
+      {env.environment === 'production' && <style dangerouslySetInnerHTML={{__html: require('./css/main.scss').toString()}}/>}
+      {env.environment === 'development' && <script defer={true} type="text/javascript" src="/webpack-bootstrap.js"></script>}
       <script defer={true} type="text/javascript" src="/js/app.js"></script>
+
+      <script type="application/ld+json" dangerouslySetInnerHTML={{__html: JSON.stringify(structuredData)}}/>
+      <meta property="og:url" content=""/>
+      <meta property="og:type" content="profile"/>
+      <meta property="og:title" content={env.title}/>
+      <meta property="og:description" content={env.description}/>
+      <meta property="og:image" content=""/>
+      <meta property="og:image:width" content="250"/>
+      <meta property="og:image:height" content="250"/>
+      <meta property="og:image:alt" content={env.description}/>
+      <meta property="og:locale" content="it"/>
+      <meta property="og:type" content="profile"/>
+      <meta property="fb:app_id" content=""/>
+      <meta property="fb:profile_id" content=""/>
+      <meta name="twitter:card" content="summary"/>
+      <meta name="twitter:site" content=""/>
+      <meta name="twitter:title" content={env.title}/>
+      <meta name="twitter:description" content={env.description}/>
+      <meta name="twitter:image" content=""/>
     </head>
     <body>
       <Provider store={store}>
-        <Router location={"/"} context={{}}>
+        <Router location={'/'} context={{}}>
           <div id="root" className="root">
             <div id="main" className="main">