+++ template/src/js/application.jsx
@@ -9,21 +9,9 @@ import ReactDOM from 'react-dom';
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
 import {Provider} from 'react-redux';
 
-// Local imports
 import {store} from './data/store';
 
-const syncServiceWorker = async function(registration){
-  // Fetch the manifest to see if version matches
-  try{
-    const response = await fetch('/manifest.json');
-    const latestVersion = await response.json();
-
-    if(latestVersion !== version)
-      registration.update().catch(console.error);
-  }catch(e){
-    // No-op
-  }
-};
+// Routes
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
@@ -38,7 +26,7 @@ document.addEventListener('DOMContentLoaded', () => {
   // Service workers
   if(navigator.serviceWorker){
     if(env.serviceWorkerEnabled && env.environment === 'production')
-      navigator.serviceWorker.register('/sw.js').then(syncServiceWorker).catch(console.error);
+      navigator.serviceWorker.register('/sw.js').catch(console.error);
     else
       navigator.serviceWorker.getRegistrations().then(registrations => registrations.map(r => r.unregister()));
   }