+++ template/src/js/data/reducers.js
@@ -1,0 +1,14 @@
+const initialState = {};
+
+export function reducer(state, {type/*, payload*/}){
+  let newState = state;
+
+  if(typeof newState === 'undefined')
+    newState = initialState;
+
+  switch(type){
+
+  }
+
+  return newState;
+}