+++ template/src/js/service-worker.js
@@ -1,44 +1,15 @@
-const activate = async function(){
-  const keys = await caches.keys();
-  const toDelete = keys.filter(k => k !== version);
+// Load Workbox
+require('../../node_modules/workbox-sw/build/importScripts/workbox-sw.prod.v2.0.1.js');
+importScripts('/js/workbox.js');
+const workbox = new WorkboxSW({clientsClaim: true});
 
-  if(!toDelete.length)
-    return Promise.resolve();
+// Cache Google Fonts
+workbox.router.registerRoute(
+  new RegExp('^(https://fonts.gstatic.com)'),
+  workbox.strategies.cacheFirst({cacheName: 'google-fonts', cacheableResponse: {statuses: [0, 200]}, networkTimeoutSeconds: 4})
+);
 
-  // Delete all previous caches
-  return Promise.all(toDelete.map(k => caches.delete(k)));
-};
-
-const fetchWithCache = async function(request){
-  if(!request.url.includes('cowtech.it')) // Don't try to cache external resources
-    return fetch(request);
-
-  // Search in the cache
-  const cache = await caches.open(version);
-  let cachedResponse = await cache.match(request);
-
-  if(cachedResponse)
-    return cachedResponse;
-
-  // Not found, make a real call
-  const fetchRequest = request.clone(); // Clone is needed in order to consume stream from both browser and cache
-  const response = await fetch(fetchRequest);
-  cachedResponse = response.clone(); // Clone is needed in order to consume stream from both browser and cache
-
-  // Request failed
-  if(!response.ok)
-    throw new Error(`Request failed with error ${response.status}`);
-
-  // Save in the cache and then return to the client
-  await cache.put(request, response);
-
-  return cachedResponse;
-};
-
-self.addEventListener('activate', event => {
-  event.waitUntil(activate());
-});
-
-self.addEventListener('fetch', event => {
-  event.respondWith(fetchWithCache(event.request));
-});
+// Application
+workbox.precache([]);
+workbox.router.registerRoute(/\/images\/(.*\/)?.*\.(png|jpg|jpeg|gif)/, workbox.strategies.cacheFirst());
+workbox.router.registerNavigationRoute('/index.html');